backend/test_requirements.txt
pytest==9.0.3
pytest-django==4.12.0